requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2==2.9.10
pygments==2.18.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-rest-swagger==2.2.0
django-simple-history-3.7.0
